feat(offline): tile removal by region

+++ packages/core/src/lib/storage/region-db.service.ts
@@ -52,20 +52,15 @@ export class RegionDBService {
     return this.dbService.getAll(this.dbName);
   }
 
-  remove(region: DBRegion): Observable<boolean> {
+  delete(region: DBRegion): Observable<boolean> {
+    if (!region) {
+      return;
+    }
+
     const dbRequest = this.dbService.deleteByKey(this.dbName, region.id);
     dbRequest.subscribe(() => {
       this.update$.next(true);
     })
-    return dbRequest;
-  }
-
-  delete(id: number): Observable<boolean> {
-    if (!id) {
-      return;
-    }
-
-    const dbRequest = this.dbService.deleteByKey(this.dbName, id);
     return dbRequest;
   }
 }
